mcp/client/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.61.15",
    "mcp>=1.3.0",
    "python-dotenv>=1.0.1",
]
